Better separation of mobile and desktop meinMenu
