+++ components/Onboarding/style.module.scss
@@ -82,6 +82,7 @@
   color: white;
   text-decoration: none;
   cursor: pointer;
+  border: 0px solid transparent;
 
   clip-path: polygon(95% 0, 100% 50%, 95% 100%, 0% 100%, 5% 50%, 0% 0%);
