+++ layout/Header/MainMenu/CustomMenuLink.tsx
@@ -11,12 +11,14 @@ type CustomMenuLinkProps = {
   entry: CustomEntry;
   currentRoute: string;
   isMobile?: boolean;
+  extraCallback?: () => void;
 };
 
 export const CustomMenuLink = ({
   entry,
   currentRoute,
   isMobile = false,
+  extraCallback,
 }: CustomMenuLinkProps): ReactElement => {
   const prefixedSlug =
     entry.slug.substring(0, 1) === '/' ? entry.slug : `/${entry.slug}`;
@@ -24,6 +26,9 @@ export const CustomMenuLink = ({
   return (
     <Link key={entry.id} href={prefixedSlug}>
       <a
+        onClick={() => {
+          if (extraCallback) extraCallback();
+        }}
         className={`${!isMobile ? 'mx-2' : ''} text-xl nowrap cursor-pointer ${
           prefixedSlug === currentRoute ? 'underline' : 'hoverUnderline'
         }`}
