+++ layout/Header/MainMenu/MenuLink.tsx
@@ -9,6 +9,7 @@ type MenuLinkProps = {
   underlineColor?: UnderlineColor;
   hoverUnderlineColor?: UnderlineColor;
   isMobile?: boolean;
+  extraCallback?: () => void;
 };
 
 export const MenuLink = ({
@@ -17,6 +18,7 @@ export const MenuLink = ({
   underlineColor = 'BLACK',
   hoverUnderlineColor = 'BLACK',
   isMobile = false,
+  extraCallback,
 }: MenuLinkProps) => {
   const prefixedSlug =
     entry.slug.substring(0, 1) === '/' ? entry.slug : `/${entry.slug}`;
@@ -24,6 +26,9 @@ export const MenuLink = ({
   return (
     <Link key={entry.id} href={prefixedSlug}>
       <a
+        onClick={() => {
+          if (extraCallback) extraCallback();
+        }}
         className={`${isMobile ? 'my-2' : 'mx-2'} text-xl nowrap ${
           prefixedSlug === currentRoute
             ? `underline${underlineColor}`
