+++ layout/Header/MainMenu/UserMenuLink.tsx
@@ -68,69 +68,3 @@ export const UserMenuLink = ({
   );
 };
 
-export const UserMenuLinkMobile = ({
-  entry,
-  currentRoute,
-}: {
-  entry: CustomEntry;
-  currentRoute: string;
-}) => {
-  const { isAuthenticated, customUserData, userId, signUserOut } =
-    useContext(AuthContext);
-  const { togglePageBuilder } = useContext(XbgeAppContext);
-  const router = useRouter();
-
-  if (!isAuthenticated) {
-    return (
-      <CustomMenuLink
-        isMobile={true}
-        entry={entry}
-        currentRoute={currentRoute}
-      />
-    );
-  }
-
-  return (
-    <>
-      {customUserData ? (
-        <>
-          <div className="items-center my-4">
-            <AvatarImage
-              className={s.loginParentAvatar}
-              user={customUserData}
-              sizes="500"
-            />
-            <span className="text-xl nowrap">{customUserData.username}</span>
-          </div>
-          <div className="mx-2">
-            <div className="my-4">
-              <CustomMenuAction
-                entry={{
-                  action: () => router.push(`/mensch/${userId}`),
-                  label: 'Zum Profil',
-                }}
-              />
-            </div>
-            {customUserData?.directus?.token && (
-              <div className="my-4">
-                <CustomMenuAction
-                  entry={{
-                    action: () => togglePageBuilder(),
-                    label: 'Page Builder',
-                  }}
-                />
-              </div>
-            )}
-            <div className="my-4">
-              <CustomMenuAction
-                entry={{ action: signUserOut, label: 'Abmelden' }}
-              />
-            </div>
-          </div>
-        </>
-      ) : (
-        <span className="text-xl nowrap ml-2">Lade...</span>
-      )}
-    </>
-  );
-};
