+++ layout/Header/MainMenu/style.module.scss
@@ -6,12 +6,14 @@
 .dropdownContent {
   display: none;
   position: absolute;
-  background-color: white;
+  text-align: right;
+  right: -2rem;
   min-width: 160px;
   // box-shadow: 0px 2px 4px 0px rgba(0, 0, 0, 0.2);
   padding: 12px 16px;
   padding-top: 0.5rem;
   z-index: 1;
+  background-color: white;
 }
 
 .dropdown:hover .dropdownContent {
