+++ layout/Header/index.tsx
@@ -2,8 +2,9 @@ import cN from 'classnames';
 import { ReactElement, useEffect, useState } from 'react';
 import { useWindowDimensions } from '../../hooks/useWindowDimensions';
 import { Menu } from '../../utils/getMenus';
-import { MainMenu, MainMenuMobile } from './MainMenu';
-import { HamburgerMenu } from './MainMenu/HamburgerMenu';
+import { MainMenu } from './MainMenu/Desktop';
+import { MainMenuMobile } from './MainMenu/Mobile';
+import { HamburgerMenu } from './MainMenu/Mobile/HamburgerMenu';
 import { PageLogo } from './PageLogo';
 import s from './style.module.scss';
 
@@ -41,7 +42,11 @@ export const Header = ({
       </div>
       {isMobile && mobileMenuActive && (
         <div className={cN(s.mobileMenu)}>
-          <MainMenuMobile mainMenu={mainMenu} currentRoute={currentRoute} />
+          <MainMenuMobile
+            mainMenu={mainMenu}
+            currentRoute={currentRoute}
+            closeMenu={() => setMobileMenuActive(false)}
+          />
         </div>
       )}
     </>
